renaming gpu4pyscf to gpu4mrh
