+++ gpu/gpu4mrh/mcscf/patch_pyscf.py
@@ -1,8 +1,8 @@
-from gpu4pyscf.mcscf import df
+from gpu4mrh.mcscf import df
 __init__ = df._ERIS.__init__
 from pyscf.mcscf.df import _ERIS
 
-from gpu4pyscf.lib.utils import patch_cpu_kernel
+from gpu4mrh.lib.utils import patch_cpu_kernel
 
 print(f'{__init__} inside CASSCF DF ERIS monkey-patched to GPU accelerated version')
