+++ gpu/gpu4mrh/patch_pyscf.py
@@ -1,0 +1,4 @@
+from gpu4mrh.df import patch_pyscf
+from gpu4mrh.gto import patch_pyscf
+from gpu4mrh.mcscf import patch_pyscf
+del patch_pyscf
