+++ gpu/mini-apps/ao2mo/runner2.py
@@ -17,7 +17,7 @@ if gpu_run:from mrh.my_pyscf.gpu import libgpu
 import pyscf
 import numpy 
 import ctypes
-if gpu_run:from gpu4pyscf import patch_pyscf
+if gpu_run:from gpu4mrh import patch_pyscf
 from mrh.tests.gpu.geometry_generator import generator
 from pyscf import gto, scf, tools, mcscf, lib
 from mrh.my_pyscf.mcscf.lasscf_async import LASSCF
