+++ gpu/mini-apps/openmp/python/main.py
@@ -3,7 +3,7 @@
 from mrh.my_pyscf.gpu import libgpu
 
 import pyscf   # -- this is contaminating a path preventing an OpenMP runtime that supports GPUs from being picked up
-from gpu4pyscf import patch_pyscf
+from gpu4mrh import patch_pyscf
 
 from pyscf import gto, scf, tools, mcscf, lib
 from mrh.my_pyscf.mcscf.lasscf_async import LASSCF
